+++ yagura-v0.35.0/cmd/yagura/main.go
@@ -60,7 +60,7 @@ import (
 
 const (
 	serviceName = "yagura"
-	version     = "0.84.0"
+	version     = "0.85.0"
 
 	// graceful shutdown 関連
 	readyDrainGrace   = 5 * time.Second